feat: display todo details when todo is clicked
