+++ src/components/ListCard.ts
@@ -1,83 +1,6 @@
 import { Todo } from '../models/todo';
 import { inlineStyles } from '../utils/styles';
-import CheckmarkIcon from './CheckmarkIcon';
-
-function ListCardTodo({ todo }: { todo: Todo }) {
-  const container = document.createElement('div');
-  container.style.cssText = inlineStyles({
-    display: 'flex',
-    padding: '0.5rem 1rem',
-    gap: '0.5rem',
-    alignItems: 'center',
-    borderRadius: '0.5rem',
-    cursor: 'pointer',
-  });
-
-  function ButtonText() {
-    return todo.complete
-      ? CheckmarkIcon({ width: 40, height: 40, fill: 'currentColor' })
-      : document.createElement('span');
-  }
-
-  const button = document.createElement('button');
-  button.appendChild(ButtonText());
-  button.type = 'button';
-  button.style.cssText = inlineStyles({
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: '1rem',
-    aspectRatio: '1',
-    borderRadius: '50%',
-    border: '1px solid #ffffff',
-    color: '#ffffff',
-  });
-
-  const span = document.createElement('span');
-  span.innerText = todo.title;
-
-  container.appendChild(button);
-  container.appendChild(span);
-
-  function handleButtonClick() {
-    // TODO: update todo in local storage
-    if (todo.complete == false) {
-      todo.complete = true;
-      button.innerHTML = '';
-      button.appendChild(ButtonText());
-
-      span.style.cssText = inlineStyles({
-        textDecoration: 'line-through',
-      });
-      return;
-    }
-
-    todo.complete = false;
-    button.innerHTML = '';
-    button.appendChild(ButtonText());
-
-    span.style.cssText = inlineStyles({
-      textDecoration: '',
-    });
-
-    return;
-  }
-
-  button.addEventListener('click', handleButtonClick);
-
-  function handleContainerMouseOver() {
-    container.style.backgroundColor = '#0093e970';
-  }
-
-  function handleContainerMouseOut() {
-    container.style.backgroundColor = '';
-  }
-
-  container.addEventListener('mouseover', handleContainerMouseOver);
-  container.addEventListener('mouseout', handleContainerMouseOut);
-
-  return container;
-}
+import ListCardTodo from './ListCardTodo';
 
 export default function ListCard({ todos }: { todos: Todo[] }) {
   const card = document.createElement('div');
@@ -129,6 +52,9 @@ export default function ListCard({ todos }: { todos: Todo[] }) {
 
     [...list.querySelectorAll('li')].forEach(todo => {
       if (todo.id === targetTodo.id) {
+        document.dispatchEvent(
+          new CustomEvent('todoClick', { detail: { todoId: todo.id } })
+        );
         todo.style.cssText = inlineStyles({
           outline: 'none',
           boxShadow: '0px 0px 4px 0px #0093e9',
