+++ src/data/context.ts
@@ -47,8 +47,21 @@ export function context() {
           _todos.push(todo);
           setItem({ key: TODOS_KEY, item: _todos });
         },
+        find: (predicate: (todo: Todo) => boolean) => {
+          return _todos.find(predicate);
+        },
         findMany: (predicate: (todo: Todo) => boolean) => {
           return _todos.filter(predicate);
+        },
+        findAndReplace: (predicate: (todo: Todo) => boolean, todo: Todo) => {
+          const index = _todos.findIndex(predicate);
+
+          if (index === -1) {
+            return;
+          }
+
+          _todos[index] = todo;
+          setItem({ key: TODOS_KEY, item: _todos });
         },
       };
     },
